Rename dashbaord storage
