+++ server/storage-jdbc-clickhouse/src/main/java/org/bithon/server/storage/jdbc/clickhouse/web/DashboardStorage.java
@@ -28,7 +28,7 @@ import org.bithon.server.storage.jdbc.clickhouse.ClickHouseStorageProviderConfig
 import org.bithon.server.storage.jdbc.clickhouse.common.SecondaryIndex;
 import org.bithon.server.storage.jdbc.clickhouse.common.TableCreator;
 import org.bithon.server.storage.jdbc.common.jooq.Tables;
-import org.bithon.server.storage.jdbc.web.DashboardJdbcStorage;
+import org.bithon.server.storage.jdbc.dashboard.DashboardJdbcStorage;
 import org.jooq.Record;
 
 import java.sql.Timestamp;
