+++ server/storage-jdbc-mysql/src/main/java/org/bithon/server/storage/jdbc/mysql/DashboardMySqlStorage.java
@@ -23,7 +23,7 @@ import com.fasterxml.jackson.annotation.OptBoolean;
 import org.bithon.server.storage.dashboard.DashboardStorageConfig;
 import org.bithon.server.storage.jdbc.JdbcStorageProviderConfiguration;
 import org.bithon.server.storage.jdbc.common.jooq.Tables;
-import org.bithon.server.storage.jdbc.web.DashboardJdbcStorage;
+import org.bithon.server.storage.jdbc.dashboard.DashboardJdbcStorage;
 import org.jooq.CreateTableElementListStep;
 import org.jooq.Field;
 import org.jooq.SQLDialect;
